+++ library/src/actions/flavor/flavor.ts
@@ -2,23 +2,31 @@ import type { BaseTransformation, SuccessDataset } from '../../types/index.ts';
 
 /**
  * Flavor symbol.
+ *
+ * @beta
  */
 export declare const FlavorSymbol: unique symbol;
 
 /**
  * Flavor name type.
+ *
+ * @beta
  */
 export type FlavorName = string | number | symbol;
 
 /**
- * Flavor type.
+ * Flavor interface.
+ *
+ * @beta
  */
 export interface Flavor<TName extends FlavorName> {
   [FlavorSymbol]?: { [TValue in TName]: TValue };
 }
 
 /**
- * Flavor action type.
+ * Flavor action interface.
+ *
+ * @beta
  */
 export interface FlavorAction<TInput, TName extends FlavorName>
   extends BaseTransformation<TInput, TInput & Flavor<TName>, never> {
@@ -42,7 +50,10 @@ export interface FlavorAction<TInput, TName extends FlavorName>
  * @param name The flavor name.
  *
  * @returns A flavor action.
+ *
+ * @beta
  */
+// @__NO_SIDE_EFFECTS__
 export function flavor<TInput, TName extends FlavorName>(
   name: TName
 ): FlavorAction<TInput, TName> {