+++ library/src/actions/index.ts
@@ -21,6 +21,7 @@ export * from './excludes/index.ts';
 export * from './filterItems/index.ts';
 export * from './findItem/index.ts';
 export * from './finite/index.ts';
+export * from './flavor/index.ts';
 export * from './graphemes/index.ts';
 export * from './gtValue/index.ts';
 export * from './hash/index.ts';